+++ src/raw/client.rs
@@ -2,13 +2,12 @@
 
 use super::requests;
 use crate::{request::KvRequest, ColumnFamily};
+use grpcio::EnvBuilder;
+use std::{sync::Arc, u32};
 use tikv_client_common::{
     security::SecurityManager, BoundRange, Config, Error, Key, KvPair, Result, Value,
 };
 use tikv_client_pd::PdRpcClient;
-
-use grpcio::EnvBuilder;
-use std::{sync::Arc, u32};
 use tikv_client_store::TikvConnect;
 
 const MAX_RAW_KV_SCAN_LIMIT: u32 = 10240;
