+++ src/transaction/lock.rs
@@ -1,12 +1,11 @@
-// use crate::requests;
-use crate::request::KvRequest;
+use crate::{request::KvRequest, transaction::requests};
+use kvproto::kvrpcpb;
+use std::{
+    collections::{HashMap, HashSet},
+    sync::Arc,
+};
 use tikv_client_common::{ErrorKind, Key, Result, Timestamp};
 use tikv_client_pd::{PdClient, RegionVerId};
-
-use crate::transaction::requests;
-use kvproto::kvrpcpb;
-use std::collections::{HashMap, HashSet};
-use std::sync::Arc;
 use tikv_client_store::KvConnect;
 
 const RESOLVE_LOCK_RETRY_LIMIT: usize = 10;
