+++ src/transaction/requests.rs
@@ -1,24 +1,14 @@
 use crate::{
     backoff::Backoff,
-    request::{
-        store_builder_stream_for_key, store_builder_stream_for_keys,
-        store_builder_stream_for_range, KvRequest,
-    },
+    request::{store_stream_for_key, store_stream_for_keys, store_stream_for_range, KvRequest},
     transaction::HasLocks,
 };
-
+use futures::{future::BoxFuture, prelude::*, stream::BoxStream};
+use kvproto::{kvrpcpb, tikvpb::TikvClient};
+use std::{mem, sync::Arc};
 use tikv_client_common::{BoundRange, Error, Key, KvPair, Result, Timestamp, Value};
 use tikv_client_pd::{PdClient, StoreBuilder};
-use tikv_client_store::{KvClient, RpcFnType, Store};
-
-use futures::future::BoxFuture;
-use futures::prelude::*;
-use futures::stream::BoxStream;
-use kvproto::kvrpcpb;
-use kvproto::tikvpb::TikvClient;
-use std::mem;
-use std::sync::Arc;
-use tikv_client_store::KvConnect;
+use tikv_client_store::{KvClient, KvConnect, RpcFnType, Store};
 
 impl KvRequest for kvrpcpb::GetRequest {
     type Result = Option<Value>;
@@ -35,12 +25,12 @@ impl KvRequest for kvrpcpb::GetRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let key = mem::take(&mut self.key).into();
-        store_builder_stream_for_key(key, pd_client)
+        store_stream_for_key(key, pd_client)
     }
 
     fn map_result(mut resp: Self::RpcResponse) -> Self::Result {
@@ -94,12 +84,12 @@ impl KvRequest for kvrpcpb::BatchGetRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let keys = mem::take(&mut self.keys);
-        store_builder_stream_for_keys(keys, pd_client)
+        store_stream_for_keys(keys, pd_client)
     }
 
     fn map_result(mut resp: Self::RpcResponse) -> Self::Result {
@@ -154,14 +144,14 @@ impl KvRequest for kvrpcpb::ScanRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let start_key = mem::take(&mut self.start_key);
         let end_key = mem::take(&mut self.end_key);
         let range = BoundRange::from((start_key, end_key));
-        store_builder_stream_for_range(range, pd_client)
+        store_stream_for_range(range, pd_client)
     }
 
     fn map_result(mut resp: Self::RpcResponse) -> Self::Result {
@@ -235,7 +225,7 @@ impl KvRequest for kvrpcpb::ResolveLockRequest {
         stream::once(future::err(region_error)).boxed()
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
@@ -245,7 +235,7 @@ impl KvRequest for kvrpcpb::ResolveLockRequest {
             .expect("ResolveLockRequest context must be given ");
         let keys = mem::take(&mut self.keys);
         pd_client
-            .store_builder_for_id(context.region_id)
+            .store_for_id(context.region_id)
             .map_ok(move |store| ((context, keys), store))
             .into_stream()
             .boxed()
@@ -291,12 +281,12 @@ impl KvRequest for kvrpcpb::CleanupRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let key = mem::take(&mut self.key).into();
-        store_builder_stream_for_key(key, pd_client)
+        store_stream_for_key(key, pd_client)
     }
 
     fn map_result(resp: Self::RpcResponse) -> Self::Result {
@@ -344,12 +334,12 @@ impl KvRequest for kvrpcpb::PrewriteRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let mutations = mem::take(&mut self.mutations);
-        store_builder_stream_for_keys(mutations, pd_client)
+        store_stream_for_keys(mutations, pd_client)
     }
 
     fn map_result(_: Self::RpcResponse) -> Self::Result {}
@@ -406,12 +396,12 @@ impl KvRequest for kvrpcpb::CommitRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let keys = mem::take(&mut self.keys);
-        store_builder_stream_for_keys(keys, pd_client)
+        store_stream_for_keys(keys, pd_client)
     }
 
     fn map_result(_: Self::RpcResponse) -> Self::Result {}
@@ -454,12 +444,12 @@ impl KvRequest for kvrpcpb::BatchRollbackRequest {
         req
     }
 
-    fn store_builder_stream<PdC: PdClient>(
+    fn store_stream<PdC: PdClient>(
         &mut self,
         pd_client: Arc<PdC>,
     ) -> BoxStream<'static, Result<(Self::KeyData, StoreBuilder)>> {
         let keys = mem::take(&mut self.keys);
-        store_builder_stream_for_keys(keys, pd_client)
+        store_stream_for_keys(keys, pd_client)
     }
 
     fn map_result(_: Self::RpcResponse) -> Self::Result {}
