+++ tikv-client-common/Cargo.toml
@@ -14,7 +14,8 @@ regex = "1"
 serde = "1.0"
 serde_derive = "1.0"
 log = "0.4"
-
+proptest = "0.9"
+proptest-derive = "0.1.0"
 
 [dependencies.prometheus]
 version = "0.8"
