+++ tikv-client-pd/Cargo.toml
@@ -14,7 +14,7 @@ tokio = { version = "0.2", features = ["sync"] }
 grpcio = { version = "0.5.0", features = [ "secure", "prost-codec" ], default-features = false }
 log = "0.4"
 
-tikv-client-common = {path = "../tikv-client-common"}
+tikv-client-common = { path = "../tikv-client-common" }
 
 [dev-dependencies]
 clap = "2.32"
